Fixed tests and DatabaseType

+++ src/main/java/org/pwr/ii/database/model/DatabaseType.java
@@ -2,15 +2,15 @@ package org.pwr.ii.database.model;
 
 public class DatabaseType {
 
-    private final int typeVoltage;
+    private final double typeVoltage;
     private final String typeName;
 
-    public DatabaseType(int typeVoltage, String typeName) {
+    public DatabaseType(double typeVoltage, String typeName) {
         this.typeVoltage = typeVoltage;
         this.typeName = typeName;
     }
 
-    public int getVoltage() {
+    public double getVoltage() {
         return typeVoltage;
     }
